requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.3
asyncer==0.0.2
attrs==24.2.0
Babel==2.11.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
Brotli==1.0.9
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chainlit==1.0.400
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
dataclasses-json==0.5.14
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
diskcache==5.6.3
distro==1.9.0
durationpy==0.9
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.108.0
flask
fastapi-socketio==0.0.10
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kubernetes==31.0.0
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-text-splitters==0.0.2
langsmith==0.1.147
Lazify==0.4.0
literalai==0.0.300
llama_cpp_python==0.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.23.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.57.1
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
overrides==7.4.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pip==24.2
platformdirs==3.10.0
ply==3.11
posthog==3.7.4
prometheus_client==0.21.0
prompt-toolkit==3.0.43
propcache==0.2.1
protobuf==5.29.1
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.15.1
PyJWT==2.10.1
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-multipart==0.0.6
python-socketio==5.11.4
pytz==2024.1
pywin32==305.1
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==25.1.2
qtconsole==5.6.0
QtPy==2.4.1
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.10.6
rsa==4.9
Send2Trash==1.8.2
setuptools==75.1.0
shellingham==1.5.4
simple-websocket==1.1.0
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.36
stack-data==0.2.0
starlette==0.32.0.post1
sympy==1.13.3
syncer==2.0.3
tenacity==8.5.0
terminado==0.17.1
tiktoken==0.8.0
tinycss2==1.2.1
tokenizers==0.20.3
tomli==2.0.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
uptrace==1.28.2
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
widgetsnbextension==4.0.13
wikipedia==1.4.0
win-inet-pton==1.1.0
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
